requirements.txt
asyncpg==0.30.0
fastapi==0.115.0
uvicorn[standard]==0.30.1
httpx==0.27.0
SQLAlchemy==2.0.32
alembic==1.13.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==8.3.2
pytest-asyncio==0.24.0
black==24.8.0
flake8==7.1.1
isort==5.13.2

pydantic~=2.11.9
PyYAML~=6.0.2